Fix build error
Previously we declared the connection pool as a naked variable which
caused it to be evaluated during the build step and therefore try to
actually connect to a running database during build which obviously
doesn't work. Instead move it into a function so that it doesn't get
evaluated at build-time but at runtime.